css-text/word-break/word-break-break-all-006.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8">
<title>word-break: break-all, spacing vowel sign</title>
<meta name="assert" content="word-break: break-all means lines may break between any two typographic letter units. A spacing vowel sign should be wrapped to the next line with its base character.">
<link rel='help' href='https://drafts.csswg.org/css-text-3/#word-break-property'>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<style type='text/css'>
.test { word-break: break-all; }
/* the CSS below is not part of the test */
.test, .ref { margin: 20px;  font: 3em/1 sans-serif; position:absolute; top: 0; width:90%;}
.ref { color:red; z-index: -100; }
</style>
</head>
<body>
<div id='instructions'>Test passes if you see no red characters.</div>
<div style="position:relative">
<div class="test" lang="hi"><div id="testdiv"><span id="testspan">हिंदी हिंदी हिंदी</span><span id="testend">←</span></div></div>
<div class="ref" lang="hi"><span>हिंदी हिंदी हिं<br/>दी</span><span id="refend">←</span></div>
</div>
<div id="log" style="margin-top: 10em;"></div>
<script>
var sentenceWidth = document.getElementById('testspan').offsetWidth
document.getElementById('testdiv').style.width = String(sentenceWidth - 5)+'px'
test(function() {
assert_true(document.getElementById('testend').offsetLeft === document.getElementById('refend').offsetLeft);
}, ' The text should wrap between the last two typographic letter units');
</script>
<!--
Notes:
A typographic unit based on extended grapheme clusters groups base characters and combining characters together.
-->
</body>
</html>
